Moved code for posts to domain specific package
