+++ src/main/java/com/varun/demo/posts/model/Post.java
@@ -1,4 +1,4 @@
-package com.varun.demo.model;
+package com.varun.demo.posts.model;
 
 
 public record Post(Integer id, String title, String content) {}
