+++ src/main/java/com/varun/demo/posts/repository/PostQueries.java
@@ -1,4 +1,4 @@
-package com.varun.demo.repository;
+package com.varun.demo.posts.repository;
 
 public class PostQueries {
